refactor: 장바구니 컨트롤러 및 서비스 리팩토링
- @AuthUser Long userId 파라미터 적용하여 JWT 기반 사용자 인증 처리
- 컨트롤러 내 todo 주석 제거 및 코드 정리
- 장바구니 삭제 시, 장바구니 소유자와 현재 로그인 유저 비교 로직 추가

+++ src/main/java/com/example/outsourcing/cart/service/CartService.java
@@ -45,11 +45,13 @@ public class CartService {
         // 유저아이디로 저장된 장바구니가 있는경우 → 기존 장바구니 가게와 requestDto가게 정보 비교 필요
         if (cartOptional.isPresent()) {
             Cart cart = cartOptional.get();
+            // TODO 가게 다르면 사용자에게 메세지 보내야함
             if (cart.getStoreId() != requestDto.getStoreId()) {
-                // TODO 가게 다르면 사용자에게 메세지 보내야함
                 // 기존 장바구니에 있는 가게 메뉴를 담을 건지
                 // 아니면 새롭게 들어온 가게 메뉴로 장바구니 담을건지
                 // 어떻게 처리하지?
+                // 04_27 일단 바로 지워버리기
+                clearCart(userId);
             }
         }
         // 기존 장바구니 있지만, 가게가 같은경우
@@ -153,7 +155,14 @@ public class CartService {
 
     }
 
-    public void deleteCartItem(Long cartId) {
+    public void deleteCartItem(Long userId, Long cartId) {
+        Cart cart = cartRepository.findById(cartId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 아이디의 장바구니가 없습니다!"));
+        // 장바구니 주인의 아이디와 현재 로그인 아이디 비교
+        if (cart.getUser().getId().equals(userId)) {
+            // 에러메세지 내용 통일 (또는 권한이 없습니다?)
+            throw new RuntimeException("다른 사용자의 장바구니에 접근할 수 없습니다!");
+        }
         cartRepository.deleteById(cartId);
     }
